+++ mapsforge-map/src/main/java/org/mapsforge/map/layer/Layers.java
@@ -1,7 +1,8 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2014 Ludwig M Brinckmann
- * Copyright 2016 devemux86
+ * Copyright 2016-2019 devemux86
+ * Copyright 2016 Andrey Novikov
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -18,10 +19,7 @@ package org.mapsforge.map.layer;
 
 import org.mapsforge.map.model.DisplayModel;
 
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.List;
-import java.util.RandomAccess;
+import java.util.*;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 /**
@@ -46,6 +44,8 @@ public class Layers implements Iterable<Layer>, RandomAccess {
     }
 
     private final DisplayModel displayModel;
+    private final Map<Integer, Integer> groupIndex;
+    private final List<Integer> groupList;
     private final List<Layer> layersList;
     private final Redrawer redrawer;
 
@@ -53,6 +53,8 @@ public class Layers implements Iterable<Layer>, RandomAccess {
         this.redrawer = redrawer;
         this.displayModel = displayModel;
 
+        this.groupIndex = new HashMap<>();
+        this.groupList = new ArrayList<>();
         this.layersList = new CopyOnWriteArrayList<>();
     }
 
@@ -116,6 +118,46 @@ public class Layers implements Iterable<Layer>, RandomAccess {
         layer.assign(this.redrawer);
         if (redraw) {
             this.redrawer.redrawLayers();
+        }
+    }
+
+    /**
+     * Adds a new layer using layer groups.
+     * <p/>
+     * Note: By default a redraw will take place afterwards.
+     * To avoid this, use {@link #add(Layer, int, boolean)}.
+     *
+     * @param layer The new layer to add
+     * @param group The layer group
+     * @see List#add(Object)
+     */
+    public synchronized void add(Layer layer, int group) {
+        add(layer, group, true);
+    }
+
+    /**
+     * Adds a new layer using layer groups.
+     *
+     * @param layer  The new layer to add
+     * @param group  The layer group
+     * @param redraw Whether the map should be redrawn after adding the layer
+     * @see List#add(Object)
+     */
+    public synchronized void add(Layer layer, int group, boolean redraw) {
+        int index = this.groupList.indexOf(group);
+        if (index < 0) {
+            throw new IllegalArgumentException("unknown layer group");
+        }
+
+        index++;
+        if (index == this.groupList.size()) {
+            add(layer, redraw);
+        } else {
+            add(this.groupIndex.get(this.groupList.get(index)), layer, redraw);
+            for (int i = index; i < this.groupList.size(); i++) {
+                group = this.groupList.get(i);
+                this.groupIndex.put(group, this.groupIndex.get(group) + 1);
+            }
         }
     }
 
@@ -185,6 +227,60 @@ public class Layers implements Iterable<Layer>, RandomAccess {
         if (redraw) {
             this.redrawer.redrawLayers();
         }
+    }
+
+    /**
+     * Adds multiple new layers using layer groups.
+     * <p/>
+     * Note: By default a redraw will take place afterwards.
+     * To avoid this, use {@link #addAll(Collection, int, boolean)}.
+     *
+     * @param layers The new layers to add
+     * @param group  The layer group
+     * @see List#addAll(Collection)
+     */
+    public synchronized void addAll(Collection<Layer> layers, int group) {
+        addAll(layers, group, true);
+    }
+
+    /**
+     * Adds multiple new layers using layer groups.
+     *
+     * @param layers The new layers to add
+     * @param group  The layer group
+     * @param redraw Whether the map should be redrawn after adding the layers
+     * @see List#addAll(Collection)
+     */
+    public synchronized void addAll(Collection<Layer> layers, int group, boolean redraw) {
+        int index = this.groupList.indexOf(group);
+        if (index < 0) {
+            throw new IllegalArgumentException("unknown layer group");
+        }
+
+        index++;
+        if (index == this.groupList.size()) {
+            addAll(layers, redraw);
+        } else {
+            addAll(this.groupIndex.get(this.groupList.get(index)), layers, redraw);
+            for (int i = index; i < this.groupList.size(); i++) {
+                group = this.groupList.get(i);
+                this.groupIndex.put(group, this.groupIndex.get(group) + layers.size());
+            }
+        }
+    }
+
+    /**
+     * Adds a new layer group.
+     *
+     * @param group The layer group
+     */
+    public synchronized void addGroup(int group) {
+        if (this.groupList.contains(group)) {
+            throw new IllegalArgumentException("group added twice");
+        }
+
+        this.groupList.add(group);
+        this.groupIndex.put(group, this.layersList.size());
     }
 
     /**
@@ -276,6 +372,15 @@ public class Layers implements Iterable<Layer>, RandomAccess {
     public synchronized Layer remove(int index, boolean redraw) {
         Layer layer = this.layersList.remove(index);
         layer.unassign();
+
+        // update layer group pointers
+        for (Integer group : this.groupIndex.keySet()) {
+            int pointer = this.groupIndex.get(group);
+            if (pointer > index) {
+                this.groupIndex.put(group, pointer - 1);
+            }
+        }
+
         if (redraw) {
             this.redrawer.redrawLayers();
         }
@@ -304,8 +409,18 @@ public class Layers implements Iterable<Layer>, RandomAccess {
      */
     public synchronized boolean remove(Layer layer, boolean redraw) {
         checkIsNull(layer);
+        int index = this.layersList.indexOf(layer);
         if (this.layersList.remove(layer)) {
             layer.unassign();
+
+            // update layer group pointers
+            for (Integer group : this.groupIndex.keySet()) {
+                int pointer = this.groupIndex.get(group);
+                if (pointer > index) {
+                    this.groupIndex.put(group, pointer - 1);
+                }
+            }
+
             if (redraw) {
                 this.redrawer.redrawLayers();
             }